refactor: Update import paths for useApi from utils to hooks for consistency
