+++ mobile/app/(auth)/registerCustomer.tsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import { View, StyleSheet, Alert, Text } from "react-native";
 import { useForm, FormProvider } from "react-hook-form";
 import { useRouter } from "expo-router";
-import { useApi } from "../utils/useApi";
+import { useApi } from "../hooks/useApi";
 import { AppTextField } from "../components/inputs/AppTextField";
 import { AppButton } from "../components/buttons/AppButton";
 import { ToolBarCustomer } from "@/app/components/customer/navigation/ToolBarCustomer";
