+++ mobile/app/(main)/admin/customers/new/index.tsx
@@ -2,7 +2,7 @@ import { View, ScrollView } from "react-native";
 import React from "react";
 import { router } from "expo-router";
 import { AdminCustomerForm } from "@/app/components/admin/customer/AdminCustomerForm";
-import { useApi } from "@/app/utils/useApi";
+import { useApi } from "@/app/hooks/useApi";
 import { ToolBarAdmin } from "@/app/components/admin/navigation/ToolBarAdmin";
 
 const AdminCustomerNew = () => {
