+++ mobile/app/(main)/admin/interventions/new/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Alert } from 'react-native';
 import { router } from 'expo-router';
 import { AdminInterventionStepper } from '@/app/components/admin/intervention/AdminInterventionStepper';
-import { useApi } from '@/app/utils/useApi';
+import { useApi } from '@/app/hooks/useApi';
 
 const AdminInterventionsNewPage = () => {
   const api = useApi();
