+++ mobile/app/(main)/customer/intervention/[id]/index.tsx
@@ -1,7 +1,7 @@
 import { useLocalSearchParams, useRouter } from "expo-router";
 import { View, Text, StyleSheet, ScrollView } from "react-native";
 import React, { useEffect, useState } from "react";
-import { useApi } from "@/app/utils/useApi";
+import { useApi } from "@/app/hooks/useApi";
 import { ToolBarCustomer } from "@/app/components/customer/navigation/ToolBarCustomer";
 
 export default function InterventionDetailPage() {
