+++ mobile/app/(main)/customer/settings/profile/index.tsx
@@ -3,7 +3,7 @@ import { View, Text, StyleSheet, ScrollView } from "react-native";
 import { AppTextField } from "@/app/components/inputs/AppTextField";
 import { useSessionContext } from "@/app/context/useSessionContext";
 import { FormProvider, useForm } from "react-hook-form";
-import { useApi } from "@/app/utils/useApi";
+import { useApi } from "@/app/hooks/useApi";
 import { AppButton } from "@/app/components/buttons/AppButton";
 import { router } from "expo-router";
 import { ToolBarCustomer } from "@/app/components/customer/navigation/ToolBarCustomer";
