+++ mobile/app/components/customer/appointment/AppointmentModalForm.tsx
@@ -10,7 +10,7 @@ import {
 import { AppButton } from "../../buttons/AppButton";
 import { FormProvider, useForm } from "react-hook-form";
 import { useSessionContext } from "@/app/context/useSessionContext";
-import { useApi } from "@/app/utils/useApi";
+import { useApi } from "@/app/hooks/useApi";
 import { AppSelectInput } from "../../inputs/AppSelectInput";
 import { Feather } from "@expo/vector-icons";
 import { AppTextField } from "../../inputs/AppTextField";
