+++ mobile/app/components/customer/appointment/TimeSlotPicker.tsx
@@ -2,7 +2,7 @@ import React, { FC, useState, useEffect } from "react";
 import { View, Text, StyleSheet, ScrollView, TouchableOpacity } from "react-native";
 import { format, addDays, startOfDay, isSameDay } from "date-fns";
 import { fr } from "date-fns/locale";
-import { useApi } from "@/app/utils/useApi";
+import { useApi } from "@/app/hooks/useApi";
 
 interface TimeSlotPickerProps {
   companyId?: number;
