+++ mobile/app/components/customer/equipment/CreateEquipmentModal.tsx
@@ -11,7 +11,7 @@ import { FormProvider, useForm } from "react-hook-form";
 import { Feather } from "@expo/vector-icons";
 import { AppTextField } from "../../inputs/AppTextField";
 import { AppSelectInput } from "../../inputs/AppSelectInput";
-import { useApi } from "@/app/utils/useApi";
+import { useApi } from "@/app/hooks/useApi";
 import { useSessionContext } from "@/app/context/useSessionContext";
 import { AppButton } from "../../buttons/AppButton";
